small visualization changes

+++ multiworld/core/image_env.py
@@ -243,7 +243,7 @@ class ImageEnv(ProxyEnv, MultitaskEnv):
 
 def normalize_image(image):
     assert image.dtype == np.uint8
-    return np.float64(image) / 255.0
+    return np.float32(image) / 255.0
 
 def unormalize_image(image):
     assert image.dtype != np.uint8
